feat(core): make `ResetHandler` an interface
