+++ packages/core/src/ResetHandlerInitializer.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tempest\Core;
+
+use Tempest\Container\Container;
+use Tempest\Container\Initializer;
+
+final readonly class ResetHandlerInitializer implements Initializer
+{
+    public function initialize(Container $container): ResetHandler
+    {
+        return $container->get(GenericResetHandler::class);
+    }
+}
